eslint, prettier, tsconfigs and vite config
