+++ services/system/AuthSig/ui/eslint.config.mjs
@@ -1,0 +1,3 @@
+import rootConfig from "../../../eslint.config.mjs";
+
+export default rootConfig;
