+++ services/system/AuthSig/ui/package.json
@@ -1,59 +1,57 @@
 {
-  "name": "@psibase/authsig-ui",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "node ../../../if-build-needed.mjs 'tsc -b --noEmit && vite build'",
-    "lint": "eslint .",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@hookform/resolvers": "^3.9.0",
-    "@psibase/common-lib": "workspace:*",
-    "@radix-ui/react-alert-dialog": "^1.1.4",
-    "@radix-ui/react-avatar": "^1.1.0",
-    "@radix-ui/react-dialog": "^1.1.4",
-    "@radix-ui/react-label": "^2.1.0",
-    "@radix-ui/react-select": "^2.1.4",
-    "@radix-ui/react-separator": "^1.1.0",
-    "@radix-ui/react-slot": "^1.1.1",
-    "@tanstack/react-query": "^5.52.0",
-    "axios": "^1.7.9",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.1",
-    "dayjs": "^1.11.13",
-    "debounce": "^2.1.0",
-    "framer-motion": "^11.3.29",
-    "lucide-react": "^0.429.0",
-    "next-themes": "^0.4.4",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-hook-form": "^7.53.0",
-    "react-qr-code": "^2.0.15",
-    "react-router-dom": "^6.26.1",
-    "sonner": "^1.7.2",
-    "tailwind-merge": "^2.5.2",
-    "tailwindcss-animate": "^1.0.7",
-    "usehooks-ts": "^3.1.0",
-    "zod": "^3.23.8"
-  },
-  "devDependencies": {
-    "@eslint/js": "^9.9.0",
-    "@types/node": "^22.5.0",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@vitejs/plugin-react": "4.2.1",
-    "autoprefixer": "^10.4.20",
-    "eslint": "^9.9.0",
-    "eslint-plugin-react-hooks": "^5.1.0-rc.0",
-    "eslint-plugin-react-refresh": "^0.4.9",
-    "globals": "^15.9.0",
-    "postcss": "^8.4.41",
-    "tailwindcss": "^3.4.10",
-    "typescript": "^5.7.3",
-    "typescript-eslint": "^8.0.1",
-    "vite": "4.5.2"
-  }
+    "name": "@psibase/authsig-ui",
+    "private": true,
+    "version": "0.0.0",
+    "scripts": {
+        "dev": "vite",
+        "build": "node ../../../if-build-needed.mjs 'tsc -b --noEmit && yarn lint && vite build'",
+        "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+        "preview": "vite preview"
+    },
+    "dependencies": {
+        "@hookform/resolvers": "^3.9.0",
+        "@psibase/common-lib": "workspace:*",
+        "@radix-ui/react-alert-dialog": "^1.1.4",
+        "@radix-ui/react-avatar": "^1.1.0",
+        "@radix-ui/react-dialog": "^1.1.4",
+        "@radix-ui/react-label": "^2.1.0",
+        "@radix-ui/react-select": "^2.1.4",
+        "@radix-ui/react-separator": "^1.1.0",
+        "@radix-ui/react-slot": "^1.1.1",
+        "@tanstack/react-query": "^5.52.0",
+        "axios": "^1.7.9",
+        "class-variance-authority": "^0.7.0",
+        "clsx": "^2.1.1",
+        "dayjs": "^1.11.13",
+        "debounce": "^2.1.0",
+        "framer-motion": "^11.3.29",
+        "lucide-react": "^0.429.0",
+        "next-themes": "^0.4.4",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
+        "react-hook-form": "^7.53.0",
+        "react-qr-code": "^2.0.15",
+        "react-router-dom": "^6.26.1",
+        "sonner": "^1.7.2",
+        "tailwind-merge": "^2.5.2",
+        "tailwindcss-animate": "^1.0.7",
+        "usehooks-ts": "^3.1.0",
+        "zod": "^3.23.8"
+    },
+    "devDependencies": {
+        "@eslint/js": "^9.9.0",
+        "@types/node": "^22.5.0",
+        "@types/react": "^18.3.3",
+        "@types/react-dom": "^18.3.0",
+        "autoprefixer": "^10.4.20",
+        "eslint": "^9.9.0",
+        "eslint-plugin-react-hooks": "^5.1.0-rc.0",
+        "eslint-plugin-react-refresh": "^0.4.9",
+        "globals": "^15.9.0",
+        "postcss": "^8.4.41",
+        "tailwindcss": "^3.4.10",
+        "typescript": "^5.7.3",
+        "typescript-eslint": "^8.0.1",
+        "vite": "^5.4.1"
+    }
 }
