+++ services/system/AuthSig/ui/tsconfig.app.json
@@ -1,35 +1,8 @@
 {
-  "compilerOptions": {
-    "target": "ES2020",
-    "useDefineForClassFields": true,
-    "lib": ["ES2020", "DOM", "DOM.Iterable"],
-    "module": "ESNext",
-    "skipLibCheck": true,
-
-    /* Bundler mode */
-    "moduleResolution": "bundler",
-    "allowImportingTsExtensions": true,
-    "isolatedModules": true,
-    "moduleDetection": "force",
-    "noEmit": true,
-    "jsx": "react-jsx",
-
-    "baseUrl": ".",
-    "paths": {
-      "@/*": ["./src/*"],
-      "/common": ["../../../user/CommonApi/common"],
-      "@psibase/common-lib": [
-        "../../../user/CommonApi/common/packages/common-lib/src"
-      ],
-      "@psibase/common-lib/*": [
-        "../../../user/CommonApi/common/packages/common-lib/src/*"
-      ]
+    "extends": "../../../tsconfig.app.json",
+    "compilerOptions": {
+        "tsBuildInfoFile": ".tmp/tsconfig.app.tsbuildinfo",
+        "baseUrl": "./"
     },
-    /* Linting */
-    "strict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noFallthroughCasesInSwitch": true
-  },
-  "include": ["src", "src/hooks/.ts"]
+    "include": ["src"]
 }
