+++ services/system/AuthSig/ui/tsconfig.node.json
@@ -1,22 +1,8 @@
 {
-  "compilerOptions": {
-    "target": "ES2022",
-    "lib": ["ES2023"],
-    "module": "ESNext",
-    "skipLibCheck": true,
-
-    /* Bundler mode */
-    "moduleResolution": "bundler",
-    "allowImportingTsExtensions": true,
-    "isolatedModules": true,
-    "moduleDetection": "force",
-    "noEmit": true,
-
-    /* Linting */
-    "strict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noFallthroughCasesInSwitch": true
-  },
-  "include": ["vite.config.ts"]
+    "extends": "../../../tsconfig.node.json",
+    "compilerOptions": {
+        "tsBuildInfoFile": ".tmp/tsconfig.node.tsbuildinfo",
+        "baseUrl": "./"
+    },
+    "include": ["vite.config.mts"]
 }
